requirements.txt
telebot~=0.0.5
pyTelegramBotAPI~=4.23.0
PyYAML~=6.0.2
pg8000~=1.31.2
fpdf~=1.7.2